requirements.txt
streamlit
google-generativeai
groq
transformers
tokenizers
python-dotenv
google-cloud-aiplatform
sentencepiece